feat!: add config system that saves mappings
Instead of inferring the type of mappings from the directory structure
of the sources folder, we now read it from a new .shulkr file.

In an upcoming release, the mappings type will no longer be tied to the
decompiler. Then, we will not be able to infer the mappings from the
sources' directory structure. This change makes the mappings detection
future-proof.

Resolves #11