revert: add `--decompiler` option
To maintain yarn compatability, we are removing the --decompiler option.
To my knowledge, yarn does not allow its users to specify a decompiler.
Until I know for sure that yarn mappings are completely useless, it's a
good idea to maintain support for backwards compatability. The default
decompiler (CFR) will be used once again.

This reverts commit 23097657fca8b3a26b0e5f7aa7a94f8b73f6af5d.